requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
apscheduler>=3.11.0
aiosqlite>=0.21.0
jinja2>=3.1.0
python-multipart>=0.0.18
pyyaml>=6.0
slack-sdk>=3.33.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
